Refactor VM and update types
Introduce a new Virtual Machine (VM) module (`src/core/vm`) with the
following submodules:
- `op`: Defines the VM's instruction set and their byte encodings.
- `scanner`: Decodes bytecode into `Op` instructions.
- `stack`: Implements a persistent stack data structure.

The VM implements the opcode semantics described previously. Some
opcodes that normally require external data (signatures, transaction
inputs, hashing) are implemented as placeholders or use ledger data
where it makes sense (`Supply`, `Height`).

Additionally, update the following types:
- `BlockMetadata.height`, `available_supply`, `locked_supply` from `u64`
  to `u32`.
- `TransactionError::InsufficientInputAmount` total input and output
  from `u64` to `u32`.
- `miner::build_mining_tx_deterministic` `max_attempts` from `u64` to
  `usize`.
- `main.build_next_block` `max_attempts` from `u64` to `usize`.

Remove the `get_tx_outpus` method from `InMemoryLedger` as it is no
longer used.

+++ src/main.rs
@@ -65,8 +65,8 @@ fn main() {
             hex::encode(prev_block.hash)
         );
 
-        // Mine the next block. We use u64::MAX attempts to mine "forever" until a solution is found.
-        match miner::build_next_block(&ledger, &prev_tx_hash, u64::MAX) {
+        // Mine the next block. We use usize::MAX attempts to mine "forever" until a solution is found.
+        match miner::build_next_block(&ledger, &prev_tx_hash, usize::MAX) {
             Some((_signing_key, new_block)) => {
                 let new_block_hash = new_block.header().hash::<Blake2s256>();
                 println!(
